uetss_bot/requirements.txt
langchain
langchain-community
langgraph
faiss-cpu
groq
python-dotenv
tiktoken
fastapi
uvicorn
sentence-transformers
groq
GroqChat